requirements.txt
tweepy==4.14.0
wikipedia==1.4.0
google-cloud-secret-manager==2.16.2